requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
sqlalchemy>=2.0.36
alembic>=1.14.0
openai>=1.60.0
python-dotenv>=1.0.1
botbuilder-core>=4.16.2
botbuilder-schema>=4.16.2
aiohttp>=3.11.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
